frontend: Refactor some imports and exports
- Combine few component files which were exported and only used in one other component
