+++ frontend/src/app/App.tsx
@@ -2,11 +2,11 @@ import {FetchBaseQueryError} from "@reduxjs/toolkit/query";
 import {Container, Footer, IconSignout, Navigation} from "hds-react";
 import React, {useEffect, useState} from "react";
 import {Link, Outlet} from "react-router-dom";
-import Notifications from "../common/components/Notifications";
-import Spinner from "../common/components/Spinner";
+import {Notifications} from "../common/components";
 import {hdsToast} from "../common/utils";
 import "../styles/index.sass";
 import {selectIsAuthenticated, selectIsAuthenticating, setIsAuthenticated} from "./authSlice";
+import {Spinner} from "./components/Spinner";
 import {useAppDispatch, useAppSelector} from "./hooks";
 import {useGetUserInfoQuery} from "./services";
 import useAuthentication from "./useAuthentication";
