+++ frontend/src/common/components/filters/index.ts
@@ -1,13 +1,5 @@
-import FilterCheckboxField from "./FilterCheckboxField";
-import FilterIntegerField from "./FilterIntegerField";
-import FilterRelatedModelComboboxField from "./FilterRelatedModelComboboxField";
-import FilterSelectField from "./FilterSelectField";
-import FilterTextInputField from "./FilterTextInputField";
-
-export {
-    FilterCheckboxField,
-    FilterIntegerField,
-    FilterRelatedModelComboboxField,
-    FilterSelectField,
-    FilterTextInputField,
-};
+export {default as FilterCheckboxField} from "./FilterCheckboxField";
+export {default as FilterIntegerField} from "./FilterIntegerField";
+export {default as FilterRelatedModelComboboxField} from "./FilterRelatedModelComboboxField";
+export {default as FilterSelectField} from "./FilterSelectField";
+export {default as FilterTextInputField} from "./FilterTextInputField";
