+++ frontend/src/features/apartment/ApartmentNewSalePage/index.ts
@@ -1,3 +1,1 @@
-import ApartmentSalePage from "./ApartmentSalePage";
-
-export default ApartmentSalePage;
+export {default} from "./ApartmentSalePage";
