+++ frontend/src/features/codes/index.ts
@@ -1,4 +1,1 @@
-import Codes from "./Codes";
-import IndicesList from "./IndicesList";
-
-export {Codes, IndicesList};
+export {default} from "./CodesPage";
